requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
azureopenai==0.0.1
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
frozenlist==1.5.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Levenshtein==0.25.1
limits==3.13.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.1.0
numpy==1.24.4
openai==1.56.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
rapidfuzz==3.9.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
scikit-learn==1.3.2
scipy==1.10.1
slowapi==0.1.9
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
threadpoolctl==3.5.0
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
wrapt==1.17.0
yarl==1.15.2
zipp==3.20.2
